feat(api): use refactored components for memory weave API
