+++ memoryweave/api/hybrid_memory_weave.py
@@ -16,7 +16,7 @@ from memoryweave.api.llm_provider import LLMProvider
 from memoryweave.api.memory_weave import MemoryWeaveAPI
 from memoryweave.components.retrieval_strategies.hybrid_fabric_strategy import HybridFabricStrategy
 from memoryweave.components.text_chunker import TextChunker
-from memoryweave.storage.hybrid_memory_store import HybridMemoryAdapter, HybridMemoryStore
+from memoryweave.storage.refactored.hybrid_store import HybridMemoryAdapter, HybridMemoryStore
 
 logger = logging.getLogger(__name__)
 
@@ -67,15 +67,13 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
         """
         # Initialize chunking components first
         self.text_chunker = TextChunker()
-        self.text_chunker.initialize(
-            {
-                "chunk_size": 300,  # Larger chunks than the full chunked version
-                "chunk_overlap": 30,  # Less overlap to save memory
-                "min_chunk_size": 50,
-                "respect_paragraphs": True,
-                "respect_sentences": True,
-            }
-        )
+        self.text_chunker.initialize({
+            "chunk_size": 300,  # Larger chunks than the full chunked version
+            "chunk_overlap": 30,  # Less overlap to save memory
+            "min_chunk_size": 50,
+            "respect_paragraphs": True,
+            "respect_sentences": True,
+        })
 
         # Initialize hybrid memory store
         self.hybrid_memory_store = HybridMemoryStore()
@@ -369,13 +367,11 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
                     chunk_text = f"{role}: {content}"
 
                     chunk_metadata = metadata.copy()
-                    chunk_metadata.update(
-                        {
-                            "chunk_index": i,
-                            "is_conversation": True,
-                            "role": role,
-                        }
-                    )
+                    chunk_metadata.update({
+                        "chunk_index": i,
+                        "is_conversation": True,
+                        "role": role,
+                    })
 
                     chunks.append({"text": chunk_text, "metadata": chunk_metadata})
 
@@ -450,13 +446,11 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
                 chunk_text = f"{role}: {content}"
 
                 chunk_metadata = metadata.copy()
-                chunk_metadata.update(
-                    {
-                        "chunk_index": i,
-                        "is_conversation": True,
-                        "role": role,
-                    }
-                )
+                chunk_metadata.update({
+                    "chunk_index": i,
+                    "is_conversation": True,
+                    "role": role,
+                })
 
                 chunks.append({"text": chunk_text, "metadata": chunk_metadata})
             return chunks
@@ -471,14 +465,12 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
         first_text = f"{first_role}: {first_content}"
 
         first_metadata = metadata.copy()
-        first_metadata.update(
-            {
-                "chunk_index": 0,
-                "is_conversation": True,
-                "role": first_role,
-                "is_first": True,
-            }
-        )
+        first_metadata.update({
+            "chunk_index": 0,
+            "is_conversation": True,
+            "role": first_role,
+            "is_first": True,
+        })
 
         chunks.append({"text": first_text, "metadata": first_metadata})
 
@@ -499,13 +491,11 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
                 )
 
                 group_metadata = metadata.copy()
-                group_metadata.update(
-                    {
-                        "chunk_index": len(chunks),
-                        "is_conversation": True,
-                        "turn_range": (i + 1, min(i + group_size, middle_count)),
-                    }
-                )
+                group_metadata.update({
+                    "chunk_index": len(chunks),
+                    "is_conversation": True,
+                    "turn_range": (i + 1, min(i + group_size, middle_count)),
+                })
 
                 chunks.append({"text": group_text, "metadata": group_metadata})
 
@@ -516,14 +506,12 @@ class HybridMemoryWeaveAPI(MemoryWeaveAPI):
         last_text = f"{last_role}: {last_content}"
 
         last_metadata = metadata.copy()
-        last_metadata.update(
-            {
-                "chunk_index": len(chunks),
-                "is_conversation": True,
-                "role": last_role,
-                "is_last": True,
-            }
-        )
+        last_metadata.update({
+            "chunk_index": len(chunks),
+            "is_conversation": True,
+            "role": last_role,
+            "is_last": True,
+        })
 
         chunks.append({"text": last_text, "metadata": last_metadata})
